快完成了

+++ database/migrations/2019_01_02_211716_create_cards_table.php
@@ -16,7 +16,7 @@ class CreateCardsTable extends Migration
         Schema::create('cards', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name')->unique()->comment('卡号');
-            $table->integer('amount')->default(0)->comment('金额');
+            $table->unsignedBigInteger('amount')->default(0)->comment('金额');
             $table->string('password')->comment('密码');
             $table->boolean('status')->default(0)->comment('封卡0:正常;1:封卡');
             $table->timestamps();
